+++ libs/app/processors/processors/index.ts
@@ -2,6 +2,7 @@ export * from './blockquote.processor';
 export * from './code.processor';
 export * from './demo.processor';
 export * from './demo-pane.processor';
+export * from './heading-anchor.processor';
 export * from './icon.processor';
 export * from './link.processor';
 export * from './playground.processor';